catching up (#1)
* Disable phpunit typehint patch on 4.3 branch

* Remove unneeded phpdocs

* Fix compatibility with PHPUnit 8

* fix merge

* [Lock] add type-hint whenever possible

* [Cache] cs fix

* [Cache] fix merge

* Remove ForwardCompatTestTrait

* Fix s-maxage=3 transient test

* Fix test compatibility with 4.x components

* Lower EOL date.

* [HttpClient] fix data loss when streaming as a PHP resource

* cleanup remaining param  and internal Intl FulLTransformer

* [Console] add missing type hints

* Improve some URLs

* [Security][Http] Add type-hints

* [Http-Kernel][5.0] Add type-hints

* [Intl][5.0] Add parameters type-hints

* [Intl] use strict comparisons

* [Validator] Allow objects implementing __toString() to be used as violation messages

[Validator] updated changelog

[Validator] updated typehint for ConstraintViolationInterface::getMessage()

[Validator] fixed spacing issue inadvertantly added in previous commit

[Validator] fixed coding standard issues

[Validator] Address feedback

[Validator] Fix coding standard violation

[Validator] update tests

[Validator] Address feedback, Round 2

[Validator] Document ConstraintViolationBuilder::__construct()  parameter

[Validator] Update changelog

[Validator] Adjust parameter documentation order

* [HttpFoundation] deprecate using $first in get and added key in all

* Fix negative DateInterval

* Fix tst patern to handle callstack with/without return typehint

* Fix deprecation in 4.4 branche

* Replace warning by isolated test

* Fix deprecation test on master

* [HttpFoundation] Clear invalid session cookie

* Add compatibility trait for PHPUnit constraint classes

* Add deprecation for method signature

* Added the missing translations for the Slovak 'sk' locale.

* Support ISO 3166-1 Alpha-3 country codes

* fix Danish translations

* Fix unitialized variable in DeprecationErrorHandler

* Fix typo

* Use PHPUnit 8.3 on Travis when possible

* Bump minimal requirements

* [HttpKernel] trim the leading backslash in the controller init

* [HttpKernel] fixed class having a leading \ in a route controller

* [DependencyInjection] improved exception message

* [EventDispatcher] wrong Request class

* [FrameworkBundle] Remove legacy route loader container

* Added translations in validator for Serbian Latin

* Added translations in validator for Serbian Cyrillic

* [Translation] Highlight invalid translation status

* [Security] Cleanup "Digest nonce has expired." translation

* Disable PHPUnit result cache on the CI

* cleanups

* Bump minimal requirements

* [HttpFoundation] some cleanups

* [HttpFoundation] fix typo

* add type-hints

* [Cache][DI] cleanup

* [Form] fix return type on FormDataCollector

* [Serializer] Encode empty objects as objects, not arrays

Allows Normalizers to return a representation of an empty object that the encoder recognizes as such.

* [WebLink] implement PSR-13 directly

* [SecurityBundle] display the correct class name on the deprecated notice

* Update UPGRADE guide of 4.3 for EventDispatcher

* Add false type to ChoiceListFactoryInterface::createView $label argument

* Add type declarations to private DefaultChoiceListFactory methods

* [TwigBridge] add missing dep

* Fix return statements

* cs fix

* [Form] Add parameter type declarations

* [Translation] XliffLintCommand: allow .xliff file extension

* [DI] deprecate support for non-object services

* [HttpKernel][FrameworkBundle] Add alternative convention for bundle directories

* [Intl] Cleanup unused language aliases entry

* [DomCrawler] Fixed CHANGELOG markup

* [TwigBridge] Replaced plain doc block copies with inheritdoc.

* [HttpFoundation] Fix deprecation message in ::isMethodSafe()

* Fixed markdown file

* [DI] fix docblocks in Container*

* Remove unused import

* Fix getMaxFilesize() returning zero

* [Intl] Explicit check

* Partially Revert "Remove trailing space in all markdown files"

This reverts commit 5a3c19846e22399f5ce43d366346dd404e2f825f.

* [FrameworkBundle] Add parameter types from HttpKernel.

* [FrameworkBundle] Add parameter types from DependencyInjection.

* [FrameworkBundle] Add parameter types from Routing.

* cs fix

* Remove superfluous phpdoc tags

* cs fix

* [ErrorHandler] Handle return types in DebugClassLoader

* Import return annotations from vendors

* [Mailer] fixed missing property assignment

* [Console] fixed a PHP notice when there is no function

* conflict with Validator < 5.0

* trigger a deprecation not a notice

* remove deprecated method

* [VarDumper] fix annotations

* [ProxyManager] fix closure binding

* [VarDumper] Add types to private methods.

* [ProxyManagerBridge] remove false positive test case

* [DI] fix dumping lazy proxies

* [VarDumper] Add parameter type declarations.

* [Serializer] Fixed docblocks and parameter names.

* [Validator] Add more parameter types.

* [Translation] Bump dependencies and apply upstream parameter types.

* [Console] Fix incomplete output mock.

* [HttpKernel] Add more parameter types.

* [ExpressionLanguage] Add more parameter types.

* [FrameworkBundle] Add parameter types from Translation.

*  [VarDumper] Remove useless variable

+++ src/Symfony/Component/HttpFoundation/BinaryFileResponse.php
@@ -11,8 +11,8 @@
 
 namespace Symfony\Component\HttpFoundation;
 
-use Symfony\Component\HttpFoundation\File\File;
 use Symfony\Component\HttpFoundation\File\Exception\FileException;
+use Symfony\Component\HttpFoundation\File\File;
 
 /**
  * BinaryFileResponse represents an HTTP response delivering a file.
@@ -31,8 +31,8 @@ class BinaryFileResponse extends Response
      * @var File
      */
     protected $file;
-    protected $offset;
-    protected $maxlen;
+    protected $offset = 0;
+    protected $maxlen = -1;
     protected $deleteFileAfterSend = false;
 
     /**
@@ -40,11 +40,11 @@ class BinaryFileResponse extends Response
      * @param int                 $status             The response status code
      * @param array               $headers            An array of response headers
      * @param bool                $public             Files are public by default
-     * @param null|string         $contentDisposition The type of Content-Disposition to set automatically with the filename
+     * @param string|null         $contentDisposition The type of Content-Disposition to set automatically with the filename
      * @param bool                $autoEtag           Whether the ETag header should be automatically set
      * @param bool                $autoLastModified   Whether the Last-Modified header should be automatically set
      */
-    public function __construct($file, int $status = 200, array $headers = array(), bool $public = true, string $contentDisposition = null, bool $autoEtag = false, bool $autoLastModified = true)
+    public function __construct($file, int $status = 200, array $headers = [], bool $public = true, string $contentDisposition = null, bool $autoEtag = false, bool $autoLastModified = true)
     {
         parent::__construct(null, $status, $headers);
 
@@ -60,13 +60,13 @@ class BinaryFileResponse extends Response
      * @param int                 $status             The response status code
      * @param array               $headers            An array of response headers
      * @param bool                $public             Files are public by default
-     * @param null|string         $contentDisposition The type of Content-Disposition to set automatically with the filename
+     * @param string|null         $contentDisposition The type of Content-Disposition to set automatically with the filename
      * @param bool                $autoEtag           Whether the ETag header should be automatically set
      * @param bool                $autoLastModified   Whether the Last-Modified header should be automatically set
      *
      * @return static
      */
-    public static function create($file = null, $status = 200, $headers = array(), $public = true, $contentDisposition = null, $autoEtag = false, $autoLastModified = true)
+    public static function create($file = null, int $status = 200, array $headers = [], bool $public = true, string $contentDisposition = null, bool $autoEtag = false, bool $autoLastModified = true)
     {
         return new static($file, $status, $headers, $public, $contentDisposition, $autoEtag, $autoLastModified);
     }
@@ -74,16 +74,13 @@ class BinaryFileResponse extends Response
     /**
      * Sets the file to stream.
      *
-     * @param \SplFileInfo|string $file               The file to stream
-     * @param string              $contentDisposition
-     * @param bool                $autoEtag
-     * @param bool                $autoLastModified
+     * @param \SplFileInfo|string $file The file to stream
      *
      * @return $this
      *
      * @throws FileException
      */
-    public function setFile($file, $contentDisposition = null, $autoEtag = false, $autoLastModified = true)
+    public function setFile($file, string $contentDisposition = null, bool $autoEtag = false, bool $autoLastModified = true)
     {
         if (!$file instanceof File) {
             if ($file instanceof \SplFileInfo) {
@@ -153,7 +150,7 @@ class BinaryFileResponse extends Response
      *
      * @return $this
      */
-    public function setContentDisposition($disposition, $filename = '', $filenameFallback = '')
+    public function setContentDisposition(string $disposition, string $filename = '', string $filenameFallback = '')
     {
         if ('' === $filename) {
             $filename = $this->file->getFilename();
@@ -165,7 +162,7 @@ class BinaryFileResponse extends Response
             for ($i = 0, $filenameLength = mb_strlen($filename, $encoding); $i < $filenameLength; ++$i) {
                 $char = mb_substr($filename, $i, 1, $encoding);
 
-                if ('%' === $char || ord($char) < 32 || ord($char) > 126) {
+                if ('%' === $char || \ord($char) < 32 || \ord($char) > 126) {
                     $filenameFallback .= '_';
                 } else {
                     $filenameFallback .= $char;
@@ -204,7 +201,7 @@ class BinaryFileResponse extends Response
 
         if (!$this->headers->has('Accept-Ranges')) {
             // Only accept ranges on safe HTTP methods
-            $this->headers->set('Accept-Ranges', $request->isMethodSafe(false) ? 'bytes' : 'none');
+            $this->headers->set('Accept-Ranges', $request->isMethodSafe() ? 'bytes' : 'none');
         }
 
         if (self::$trustXSendfileTypeHeader && $request->headers->has('X-Sendfile-Type')) {
@@ -219,22 +216,27 @@ class BinaryFileResponse extends Response
                 // Do X-Accel-Mapping substitutions.
                 // @link http://wiki.nginx.org/X-accel#X-Accel-Redirect
                 $parts = HeaderUtils::split($request->headers->get('X-Accel-Mapping', ''), ',=');
-                $mappings = HeaderUtils::combine($parts);
-                foreach ($mappings as $pathPrefix => $location) {
-                    if (substr($path, 0, strlen($pathPrefix)) === $pathPrefix) {
-                        $path = $location.substr($path, strlen($pathPrefix));
+                foreach ($parts as $part) {
+                    list($pathPrefix, $location) = $part;
+                    if (substr($path, 0, \strlen($pathPrefix)) === $pathPrefix) {
+                        $path = $location.substr($path, \strlen($pathPrefix));
+                        // Only set X-Accel-Redirect header if a valid URI can be produced
+                        // as nginx does not serve arbitrary file paths.
+                        $this->headers->set($type, $path);
+                        $this->maxlen = 0;
                         break;
                     }
                 }
+            } else {
+                $this->headers->set($type, $path);
+                $this->maxlen = 0;
             }
-            $this->headers->set($type, $path);
-            $this->maxlen = 0;
         } elseif ($request->headers->has('Range')) {
             // Process the range headers.
             if (!$request->headers->has('If-Range') || $this->hasValidIfRangeHeader($request->headers->get('If-Range'))) {
                 $range = $request->headers->get('Range');
 
-                list($start, $end) = explode('-', substr($range, 6), 2) + array(0);
+                list($start, $end) = explode('-', substr($range, 6), 2) + [0];
 
                 $end = ('' === $end) ? $fileSize - 1 : (int) $end;
 
@@ -264,7 +266,7 @@ class BinaryFileResponse extends Response
         return $this;
     }
 
-    private function hasValidIfRangeHeader($header)
+    private function hasValidIfRangeHeader(?string $header): bool
     {
         if ($this->getEtag() === $header) {
             return true;
@@ -300,7 +302,7 @@ class BinaryFileResponse extends Response
         fclose($out);
         fclose($file);
 
-        if ($this->deleteFileAfterSend) {
+        if ($this->deleteFileAfterSend && file_exists($this->file->getPathname())) {
             unlink($this->file->getPathname());
         }
 
@@ -312,7 +314,7 @@ class BinaryFileResponse extends Response
      *
      * @throws \LogicException when the content is not null
      */
-    public function setContent($content)
+    public function setContent(?string $content)
     {
         if (null !== $content) {
             throw new \LogicException('The content cannot be set on a BinaryFileResponse instance.');
@@ -341,11 +343,9 @@ class BinaryFileResponse extends Response
      * If this is set to true, the file will be unlinked after the request is send
      * Note: If the X-Sendfile header is used, the deleteFileAfterSend setting will not be used.
      *
-     * @param bool $shouldDelete
-     *
      * @return $this
      */
-    public function deleteFileAfterSend($shouldDelete = true)
+    public function deleteFileAfterSend(bool $shouldDelete = true)
     {
         $this->deleteFileAfterSend = $shouldDelete;
